fix(hypha): resolve clippy warnings in tests
- Add #[allow(clippy::disallowed_types)] to test modules using std::sync::Mutex
- Replace assert_eq!(x, bool) with assert!(x) / assert!(!x) in cell_set tests
- Remove unnecessary u64 cast in concurrency tests

+++ hypha/src/tests/concurrency.rs
@@ -31,7 +31,7 @@ fn test_concurrent_writers() {
 
     // Cell should have some value (last write wins, but we don't know which)
     let final_value = cell.get();
-    assert!(final_value < (num_threads * writes_per_thread) as u64);
+    assert!(final_value < num_threads * writes_per_thread);
 }
 
 #[test]
